further consolidation of types
