+++ types.ts
@@ -1,3 +1,4 @@
+// BASE DRAWING/BRUSH TYPES
 export type BrushShape = "line" | "circle" | "square" | "text";
 
 export interface Paintbrush {
@@ -110,6 +111,7 @@ export type MirrorModeParams = SetupShapeReturnValues & {
   fn: p5["line"] | p5["circle"] | p5["square"] | p5["text"];
 };
 
+// WEBSOCKET TYPES
 export interface DrawUpdate
   extends Omit<Paintbrush, "fillColor" | "strokeColor"> {
   prevX: number;
@@ -152,3 +154,27 @@ export interface Connections {
 }
 
 export type Payload = LeanDrawUpdate | Messages | Connections;
+
+// CUSTOM p5 TYPES
+export type p5GuiInstance = {
+  id: string;
+  addObject: (params: GuiValues | LfoValues) => void;
+  setPosition: (x?: number, y?: number) => void;
+  collapse: () => void;
+};
+
+export type PaintWithFriends = p5 & {
+  initCanvas: (serializedCanvas: string) => void;
+  initLastCoords: () => void;
+  setLastCoords: () => void;
+  paint: () => void;
+  createGui: (name: string, p5: p5) => p5GuiInstance;
+};
+
+// SERVER TYPES
+export type ImageCacheEntry = { data: null | Buffer; shouldFetch: boolean };
+
+export type ImageCache = {
+  thumbnail: ImageCacheEntry;
+  image: ImageCacheEntry;
+};
